Fix Bugs
updateuserinfo now returns the url of the profile picture.

Change #1
----------------

On the login screen, please link "Find a RhinoFit Gym near you" to the
following page:

https://my.rhinofit.ca/findrhinofitgym

Have this link open in Safari or whatever their default browser is.

Change #2
----------------

Please change the font on the login screen for "RhinoFit" to Optima.

Change #3
----------------

On the log in screen, can you try making the background image this
attached image? kettlebell.png. Let's try it out. You may have to make
the text white in color so it shows up better.

Change #4
----------------

Track WOD is currently reversed.

canedit = 0 means that the user cannot edit the WOD or the Description

canedit = 1 means that they can edit these things

For example, log in as james.brown@rhinofit.ca pass 1234 and track the
WOD on Dec 11, 9am. The user can edit both the WOD and the title- but
canedit = 0, so they should not be able to.. see screenshot.

Go to "My Classes", Dec 12 - click TRACK WOD on the 6am Bootcamp.
Canedit = 1 here, so the user should be able to edit all 3 text areas.
See screenshot.

Change #5
----------------

When I click "Track WOD" from the Track WOD screen it calls the API
call getwodinfo instead of trackwod.

Then I got stuck on the "You have not tracked any WODs" - there was no
menu there any more to change my location.

Change #6
----------------

Log in as james.brown@rhinofit.ca, go to "My Classes", Dec 11 and click
"Mark Attended" on the 02:00pm class. The API sends an error:

"Unfortunately, you cannot attend this class....."

Please show this error in a dialog.

Change #7
----------------

Log in as james.brown@rhinofit.ca, go to "My Classes", Dec 11. The
times show up as

09:00AM
02:00PM

Please remove the leading 0 so they show up as:

9:00AM
2:00PM