Fix Bugs
Change #1
---------------
The log in screen - the text in the email and password is now white, so
it doesn't show up. Please change back to black.

Change #2
---------------
On the log in screen, please put back the text "Find a RhinoFit gym
near you" - but link it to https://my.rhinofit.ca/findrhinofitgym.

Change #3
---------------

Still same the problem as before - when I click "Track WOD" from the
"Track WOD" screen - it calls getwodinfo - it needs to call trackwod
API call.