Fix Bugs
If you attend a class with the iPhone 4S - the "TRACK WOD" text doesn't
show up fully on the screen, screenshot attached.

I am also getting an NSException when viewing the classes for December
15th under the james.brown@rhinofit.ca account. Log attached.

I also found that on the Track WOD screen that I could not scroll down
to where the "Track WOD" button was.