Fixed Bugs
Update benchmark - when I click on update, there's no way to insert a
colon for the time, for example "Fran". If we can't allow the : to be
inserted, let's have 2 separate fields, one for minutes and one for
seconds.

iPhone 5C - "TRACK WOD" is over run with text, see screen shot. Let's
put it on a new line under "You have attended this class" so it doesn't
interfere.

If there are no wall posts - it says "There are no walls" - please
change this to "There are no wall posts at this time, post one today!".

iPhone 5C - "Post a Wall Message" - The screen starts at the bottom
where the button is, and I can't scroll up to the top of the page.
Screenshot attached.

Two other things:

1) Please remove the auto-correct from the email field on the login
screen.

2) What's your IP address? I will add you to the list of verified IPs
so I can get you to log in as dropboxcrossfit@gmail.com. Once you log
in as this address the app crashes.