Fixing bug
When I go to January 13 and mark a class as attended, the date passed
to the API call makeattendance is the current date, 2015-01-14 instead
of 2014-01-13.

This happens for each date, if I go back it the past, it always send
the current date, it needs to send the date that I am currently
viewing, e.g. Dec 20th or Jan 1st.